Projects/Grammar and Speller Checker/requirements.txt
Flask==2.3.3
textblob==0.17.1
cloudscraper==1.2.71
requests==2.31.0
python-docx==0.8.11
PyPDF2==3.0.1
werkzeug==2.3.7